requirements.txt
streamlit
GitPython
pandas
plotly
PyYAML